refactor to be less WET
